// src/cards/Enums.ts
export enum CardNameString {
  ace = '1',
  two = '2',
  three = '3',
  four = '4',
  five = '5',
  six = '6',
  seven = '7',
  eight = '8',
  nine = '9',
  ten = '10',
  jack = 'jack',
  queen = 'queen',
  king = 'king',
  joker = 'joker',
}

export enum CardName {
  ace,
  two,
  three,
  four,
  five,
  six,
  seven,
  eight,
  nine,
  ten,
  jack,
  queen,
  king,
  joker,
}

export enum Suit {
  heart,
  spade,
  diamond,
  club,
}

export enum SuitTranslation {
  heart = 'Copas',
  spade = 'Espadas',
  diamond = 'Ouros',
  club = 'Paus',
}

export enum CardNameTranslation {
  ace = 'Ás',
  two = 'Dois',
  three = 'Três',
  four = 'Quatro',
  five = 'Cinco',
  six = 'Seis',
  seven = 'Sete',
  eight = 'Oito',
  nine = 'Nove',
  ten = 'Dez',
  jack = 'Valete',
  queen = 'Dama',
  king = 'Rei',
  joker = 'Joker',
}